# server/api/api/openapi.yaml
openapi: 3.0.3
info:
  title: CodeBattle API
  version: 1.0.0
servers:
- url: https
paths:
  /users/me:
    get:
      operationId: getMe
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ok
      security:
      - cookieAuth: []
      tags:
      - users
  /contests:
    get:
      operationId: getContests
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/getContests_200_response_inner'
                type: array
          description: ok
      tags:
      - contests
  /contests/{contestId}:
    get:
      operationId: getContestById
      parameters:
      - explode: false
        in: path
        name: contestId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contest'
          description: ok
      tags:
      - contests
  /contests/{contestId}/matches:
    put:
      operationId: putMatch
      parameters:
      - explode: false
        in: path
        name: contestId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
          description: ok
      security:
      - cookieAuth: []
      tags:
      - contests
  /entries:
    post:
      operationId: postEntry
      parameters: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
          description: ok
      security:
      - cookieAuth: []
      tags:
      - entries
  /entries/{entryId}:
    put:
      operationId: putEntry
      parameters:
      - explode: false
        in: path
        name: entryId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
          description: ok
      security:
      - cookieAuth: []
      tags:
      - entries
components:
  schemas:
    User:
      example:
        name: tester
        icon: https://gravatar.com/avatar/sample
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          type: string
        name:
          example: tester
          type: string
        icon:
          example: https://gravatar.com/avatar/sample
          type: string
      required:
      - id
      - name
      type: object
    Contest:
      example:
        owner:
          name: tester
          icon: https://gravatar.com/avatar/sample
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: Sample Game
        description: description
        recentMatches:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          entries:
          - null
          - null
          - null
          - null
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - createdAt: 2000-01-23T04:56:07.000+00:00
          entries:
          - null
          - null
          - null
          - null
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        ranking:
        - score: 800
          name: bot example
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          repository: http://example.com/aeiou
          error: error
          user:
            name: tester
            icon: https://gravatar.com/avatar/sample
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: processing
        - score: 800
          name: bot example
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          repository: http://example.com/aeiou
          error: error
          user:
            name: tester
            icon: https://gravatar.com/avatar/sample
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: processing
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          type: string
        owner:
          $ref: '#/components/schemas/User'
        name:
          example: Sample Game
          type: string
        description:
          type: string
        recentMatches:
          items:
            $ref: '#/components/schemas/Match'
          type: array
        ranking:
          items:
            $ref: '#/components/schemas/Entry'
          type: array
      required:
      - id
      - name
      - owner
      type: object
    Match:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        entries:
        - null
        - null
        - null
        - null
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          type: string
        createdAt:
          format: date-time
          type: string
        entries:
          items:
            $ref: '#/components/schemas/MatchEntry'
          maxItems: 4
          minItems: 2
          type: array
      required:
      - entries
      - id
      type: object
    MatchEntry:
      allOf:
      - $ref: '#/components/schemas/Entry'
      - $ref: '#/components/schemas/MatchEntry_allOf'
      required:
      - beforeScore
      type: object
    Entry:
      example:
        score: 800
        name: bot example
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        repository: http://example.com/aeiou
        error: error
        user:
          name: tester
          icon: https://gravatar.com/avatar/sample
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: processing
      properties:
        id:
          format: uuid
          type: string
        name:
          example: bot example
          maxLength: 20
          type: string
        user:
          $ref: '#/components/schemas/User'
        score:
          maximum: 9999
          minimum: 0
          type: integer
        repository:
          format: url
          type: string
        status:
          enum:
          - processing
          - registered
          - error
          type: string
        error:
          type: string
      required:
      - id
      - name
      - score
      - user
      type: object
    getContests_200_response_inner:
      example:
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
      required:
      - id
      - name
      type: object
    MatchEntry_allOf:
      properties:
        rank:
          maximum: 4
          minimum: 1
          type: integer
        beforeScore:
          maximum: 9999
          minimum: 0
          type: integer
        afterScore:
          maximum: 9999
          minimum: 0
          type: integer
  securitySchemes:
    cookieAuth:
      in: cookie
      name: JWT
      type: apiKey