# server/api/go/model_contest.go
/*
 * CodeBattle API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Contest struct {

	Id string `json:"id"`

	Owner User `json:"owner"`

	Name string `json:"name"`

	Description string `json:"description,omitempty"`

	RecentMatches []Match `json:"recentMatches,omitempty"`

	Ranking []Entry `json:"ranking,omitempty"`
}