# server/api/go/model_match_entry.go
/*
 * CodeBattle API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type MatchEntry struct {

	Id string `json:"id"`

	Name string `json:"name"`

	User User `json:"user"`

	Score int32 `json:"score"`

	Repository string `json:"repository,omitempty"`

	Status string `json:"status,omitempty"`

	Error string `json:"error,omitempty"`

	Rank int32 `json:"rank,omitempty"`

	BeforeScore int32 `json:"beforeScore"`

	AfterScore int32 `json:"afterScore,omitempty"`
}