# server/api/go/model_user.go
/*
 * CodeBattle API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type User struct {

	Id string `json:"id"`

	Name string `json:"name"`

	Icon string `json:"icon,omitempty"`
}